chore: clean up globals
