+++ packages/@eventual/core-runtime/src/providers/workflow-provider.ts
@@ -1,4 +1,4 @@
-import { Workflow } from "@eventual/core";
+import type { Workflow } from "@eventual/core";
 import { workflows, ServiceSpec } from "@eventual/core/internal";
 
 export interface WorkflowProvider extends WorkflowSpecProvider {
@@ -19,9 +19,11 @@ export class GlobalWorkflowProvider implements WorkflowProvider {
   public lookupWorkflow(workflowName: string): Workflow | undefined {
     return workflows().get(workflowName);
   }
+
   public getWorkflowNames(): string[] {
     return [...workflows().keys()];
   }
+
   public workflowExists(workflowName: string): boolean {
     return !!this.lookupWorkflow(workflowName);
   }
