+++ packages/@eventual/core/src/internal/task.ts
@@ -8,15 +8,34 @@ import type {
 } from "../task.js";
 import type { SourceLocation } from "./service-spec.js";
 
-export const AsyncTokenSymbol = /* @__PURE__ */ Symbol.for(
-  "eventual:AsyncToken"
-);
+/**
+ * Globals that may be overridden by the core-runtime. See matching core-runtime file to understand
+ * the specific behavior.
+ *
+ * In this case, we'll provide a default no-op hook function.
+ * When someone uses the enterEventualCallHookScope in runtime, the getEventualCallHook function
+ * will be overridden to return that hook (based on async scope.)
+ */
+declare global {
+  export function getEventualTaskRuntimeContext(): TaskRuntimeContext;
+}
 
 export interface TaskRuntimeContext {
   execution: TaskExecutionContext;
   invocation: TaskInvocationContext;
   service: ServiceContext;
 }
+
+// default implementation of getEventualTaskRuntimeContext that throws.
+// to be overridden by the core-runtime.
+// only set if it was not set before.
+globalThis.getEventualTaskRuntimeContext ??= () => {
+  throw new Error("Eventual task context has not been registered yet.");
+};
+
+export const AsyncTokenSymbol = /* @__PURE__ */ Symbol.for(
+  "eventual:AsyncToken"
+);
 
 export type TaskInput<A extends Task<any, any>> = A extends Task<
   string,
