+++ packages/@eventual/core/src/task.ts
@@ -8,7 +8,7 @@ import type {
   SendTaskSuccessRequest,
 } from "./internal/eventual-service.js";
 import { isTaskWorker } from "./internal/flags.js";
-import { getServiceClient, getTaskContext, tasks } from "./internal/global.js";
+import { getServiceClient, tasks } from "./internal/global.js";
 import { isDurationSchedule, isTimeSchedule } from "./internal/schedule.js";
 import { SourceLocation, isSourceLocation } from "./internal/service-spec.js";
 import { AsyncTokenSymbol, TaskSpec } from "./internal/task.js";
@@ -214,7 +214,7 @@ export async function asyncResult<Output = any>(
   if (!isTaskWorker()) {
     throw new Error("asyncResult can only be called from within a task.");
   }
-  const taskContext = await getTaskContext();
+  const taskContext = getEventualTaskRuntimeContext();
   if (!taskContext) {
     throw new Error(
       "Task context has not been set yet, asyncResult can only be used from within a task."
@@ -288,7 +288,7 @@ export function task<Name extends string, Input = any, Output = any>(
         heartbeat: options?.heartbeatTimeout ?? opts?.heartbeatTimeout,
       }),
       async () => {
-        const runtimeContext = await getTaskContext();
+        const runtimeContext = getEventualTaskRuntimeContext();
         const context: TaskContext = {
           task: {
             name,
